Simplify Azure deployment configuration
- Remove complex IISNode setup
- Add simple server.js for Next.js
- Remove turbopack from production build
- Add Node.js version requirement
- Simplify web.config
