--- pathfinder-app/app.js
@@ -1,19 +1,0 @@
-const { createServer } = require('http')
-const { parse } = require('url')
-const next = require('next')
-
-const port = process.env.PORT || 3000
-const dev = process.env.NODE_ENV !== 'production'
-
-const app = next({ dev })
-const handle = app.getRequestHandler()
-
-app.prepare().then(() => {
-  createServer((req, res) => {
-    const parsedUrl = parse(req.url, true)
-    handle(req, res, parsedUrl)
-  }).listen(port, (err) => {
-    if (err) throw err
-    console.log(`> Ready on http://localhost:${port}`)
-  })
-})
