+++ pathfinder-app/next.config.js
@@ -8,7 +8,6 @@ const nextConfig = {
     // Ignore TypeScript errors during builds for deployment
     ignoreBuildErrors: true,
   },
-  output: 'standalone',
 }
 
 module.exports = nextConfig
