+++ pathfinder-app/web.config
@@ -2,43 +2,15 @@
 <configuration>
   <system.webServer>
     <handlers>
-      <add name="iisnode" path="app.js" verb="*" modules="iisnode"/>
+      <add name="iisnode" path="server.js" verb="*" modules="iisnode"/>
     </handlers>
-    
     <rewrite>
       <rules>
-        <rule name="NodeInspector" patternSyntax="ECMAScript" stopProcessing="true">
-          <match url="^app.js\/debug[\/]?" />
-        </rule>
-        
-        <rule name="StaticContent">
-          <action type="Rewrite" url="public{REQUEST_URI}"/>
-        </rule>
-        
         <rule name="DynamicContent">
-          <conditions>
-            <add input="{REQUEST_FILENAME}" matchType="IsFile" negate="True"/>
-          </conditions>
-          <action type="Rewrite" url="app.js"/>
+          <match url="/*" />
+          <action type="Rewrite" url="server.js"/>
         </rule>
       </rules>
     </rewrite>
-    
-    <security>
-      <requestFiltering>
-        <hiddenSegments>
-          <remove segment="bin"/>
-        </hiddenSegments>
-      </requestFiltering>
-    </security>
-    
-    <httpErrors existingResponse="PassThrough" />
-    
-    <iisnode 
-      node_env="production"
-      nodeProcessCommandLine="node"
-      watchedFiles="web.config;*.js"
-      loggingEnabled="true"
-      logDirectoryNameSuffix="logs" />
   </system.webServer>
 </configuration>
